add fecth news function
